split out flatten
